pyproject.toml
[project]
name = "hia-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "filetype>=1.2.0",
    "gotrue>=2.12.4",
    "langchain-openai>=1.0.1",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pyjwt>=2.10.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "supabase>=2.22.2",
]
